+++ packages/components/textarea/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @yamada-ui/textarea
 
+## 1.1.6
+
+### Patch Changes
+
+- [#954](https://github.com/yamada-ui/yamada-ui/pull/954) [`b3df069`](https://github.com/yamada-ui/yamada-ui/commit/b3df0693c199c8809bc00083aea4b0ac26c8a52c) Thanks [@taroj1205](https://github.com/taroj1205)! - Make `useAutosize` function use rows instead of height.
+
+- Updated dependencies [[`41538f5`](https://github.com/yamada-ui/yamada-ui/commit/41538f506e383f9cfa7095f0eb0f6d7f7856ca6b), [`68ad161`](https://github.com/yamada-ui/yamada-ui/commit/68ad161f60f01708ef05dc18ce8ec4dee0cdf43a), [`575220a`](https://github.com/yamada-ui/yamada-ui/commit/575220abd642948e10e71fecbd732934805864e6)]:
+  - @yamada-ui/core@1.4.4
+  - @yamada-ui/form-control@1.0.19
+
 ## 1.1.5
 
 ### Patch Changes